requirements.txt
ipython
langchain_core
langchain_google_genai
Pillow
protobuf
python-dotenv
streamlit
